+++ src/node-transformers/converting-transformers/ObjectExpressionTransformer.ts
@@ -39,7 +39,7 @@ export class ObjectExpressionTransformer extends AbstractNodeTransformer {
      */
     public getVisitor (transformationStage: TransformationStage): IVisitor | null {
         switch (transformationStage) {
-            case TransformationStage.Obfuscating:
+            case TransformationStage.Converting:
                 return {
                     enter: (node: ESTree.Node, parentNode: ESTree.Node | null) => {
                         if (parentNode && NodeGuards.isObjectExpressionNode(node)) {