+++ test/functional-tests/node-transformers/obfuscating-transformers/class-declaration-transformer/ClassDeclarationTransformer.spec.ts
@@ -123,5 +123,56 @@ describe('ClassDeclarationTransformer', () => {
                 assert.match(obfuscatedCode, variableDeclarationsRegExp);
             });
         });
+
+        describe('Variant #5: named export', () => {
+            const namedExportRegExp: RegExp = /export class Foo *{}/;
+
+            let obfuscatedCode: string;
+
+            before(() => {
+                const code: string = readFileAsString(__dirname + '/fixtures/named-export.js');
+                const obfuscationResult: IObfuscationResult = JavaScriptObfuscator.obfuscate(
+                    code,
+                    {
+                        ...NO_ADDITIONAL_NODES_PRESET,
+                        renameGlobals: true
+                    }
+                );
+
+                obfuscatedCode = obfuscationResult.getObfuscatedCode();
+            });
+
+            it('shouldn\'t transform identifiers in named export', () => {
+                assert.match(obfuscatedCode, namedExportRegExp);
+            });
+        });
+
+        describe('Variant #6: default export', () => {
+            const classDeclarationRegExp: RegExp = /class _0x[a-f0-9]{4,6} *{}/;
+            const defaultExportRegExp: RegExp = /export default _0x[a-f0-9]{4,6};/;
+
+            let obfuscatedCode: string;
+
+            before(() => {
+                const code: string = readFileAsString(__dirname + '/fixtures/default-export.js');
+                const obfuscationResult: IObfuscationResult = JavaScriptObfuscator.obfuscate(
+                    code,
+                    {
+                        ...NO_ADDITIONAL_NODES_PRESET,
+                        renameGlobals: true
+                    }
+                );
+
+                obfuscatedCode = obfuscationResult.getObfuscatedCode();
+            });
+
+            it('Match #1: should transform identifiers in variable declaration', () => {
+                assert.match(obfuscatedCode, classDeclarationRegExp);
+            });
+
+            it('Match #2: should transform identifiers in default export', () => {
+                assert.match(obfuscatedCode, defaultExportRegExp);
+            });
+        });
     });
 });